+++ package.json
@@ -1,44 +1,4 @@
 {
-    "name": "turing-labook8",
-    "version": "1.0.0",
-    "description": "Repositório do projeto Labook",
-    "main": "index.js",
-    "scripts": {
-        "start": "node ./build/index.js",
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "start:dev": "ts-node-dev ./src/index.ts"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/future4code/turing-labook8.git"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/future4code/turing-labook8/issues"
-    },
-    "homepage": "https://github.com/future4code/turing-labook8#readme",
-    "dependencies": {
-        "@types/bcrypt": "^3.0.0",
-        "@types/bcryptjs": "^2.4.2",
-        "@types/express": "^4.17.0",
-        "@types/jsonwebtoken": "^8.5.0",
-        "@types/knex": "^0.16.1",
-        "@types/node": "^14.11.2",
-        "@types/uuid": "^8.3.0",
-        "bcrypt": "^5.0.0",
-        "bcryptjs": "^2.4.3",
-        "dotenv": "^8.2.0",
-        "express": "^4.17.0",
-        "jsonwebtoken": "^8.5.1",
-        "knex": "^0.21.5",
-        "mysql": "^2.18.1",
-        "ts-node": "^9.0.0",
-        "ts-node-dev": "^1.0.0-pre.63",
-        "typescript": "^4.0.3",
-        "uuid": "^8.3.0"
-    }
   "name": "turing-labook8",
   "version": "1.0.0",
   "description": "Repositório do projeto Labook",
